+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/catalog/_wait.py
@@ -20,7 +20,7 @@ class Wait(AAZWaitCommand):
 
     _aaz_info = {
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/catalogs/{}", "2024-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/catalogs/{}", "2025-04-01-preview"],
         ]
     }
 
@@ -136,7 +136,7 @@ class Wait(AAZWaitCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-10-01-preview",
+                    "api-version", "2025-04-01-preview",
                     required=True,
                 ),
             }
@@ -205,6 +205,7 @@ class Wait(AAZWaitCommand):
             )
             properties.last_sync_stats = AAZObjectType(
                 serialized_name="lastSyncStats",
+                flags={"read_only": True},
             )
             properties.last_sync_time = AAZStrType(
                 serialized_name="lastSyncTime",