+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/catalog_task/_show.py
@@ -22,9 +22,9 @@ class Show(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-10-01-preview",
+        "version": "2025-04-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/catalogs/{}/tasks/{}", "2024-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.devcenter/devcenters/{}/catalogs/{}/tasks/{}", "2025-04-01-preview"],
         ]
     }
 
@@ -155,7 +155,7 @@ class Show(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-10-01-preview",
+                    "api-version", "2025-04-01-preview",
                     required=True,
                 ),
             }
@@ -214,6 +214,7 @@ class Show(AAZCommand):
             )
             properties.validation_status = AAZStrType(
                 serialized_name="validationStatus",
+                flags={"read_only": True},
             )
 
             inputs = cls._schema_on_200.properties.inputs