+++ src/devcenter/azext_devcenter/aaz/latest/devcenter/admin/sku/_list.py
@@ -22,9 +22,9 @@ class List(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2024-10-01-preview",
+        "version": "2025-04-01-preview",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.devcenter/skus", "2024-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.devcenter/skus", "2025-04-01-preview"],
         ]
     }
 
@@ -103,7 +103,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2024-10-01-preview",
+                    "api-version", "2025-04-01-preview",
                     required=True,
                 ),
             }